# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vehicles/": {
      "get": {
        "tags": [
          "vehicles"
        ],
        "summary": "Get Vehicles",
        "operationId": "get_vehicles_vehicles__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "vehicles"
        ],
        "summary": "Add Vehicle",
        "operationId": "add_vehicle_vehicles__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddVehicleRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/vehicles/{numberplate}": {
      "delete": {
        "tags": [
          "vehicles"
        ],
        "summary": "Delete Vehicle",
        "operationId": "delete_vehicle_vehicles__numberplate__delete",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "numberplate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Numberplate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buildings/": {
      "get": {
        "tags": [
          "buildings"
        ],
        "summary": "Get Buildings",
        "operationId": "get_buildings_buildings__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "buildings"
        ],
        "summary": "Add Building",
        "operationId": "add_building_buildings__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBuildingRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/buildings/{building_id}/floors": {
      "post": {
        "tags": [
          "buildings"
        ],
        "summary": "Add Floor",
        "operationId": "add_floor_buildings__building_id__floors_post",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Building Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddFloorRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "buildings"
        ],
        "summary": "Get Floors",
        "operationId": "get_floors_buildings__building_id__floors_get",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Building Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buildings/{building_id}/floors/{floor_id}/slots": {
      "get": {
        "tags": [
          "buildings"
        ],
        "summary": "Get Slots",
        "operationId": "get_slots_buildings__building_id__floors__floor_id__slots_get",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Building Id"
            }
          },
          {
            "name": "floor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Floor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buildings/{building_id}/offices": {
      "post": {
        "tags": [
          "buildings",
          "offices"
        ],
        "summary": "Add Office",
        "operationId": "add_office_buildings__building_id__offices_post",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Building Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOfficeRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/buildings/{building_id}/offices/{office_id}": {
      "delete": {
        "tags": [
          "buildings",
          "offices"
        ],
        "summary": "Delete Office",
        "operationId": "delete_office_buildings__building_id__offices__office_id__delete",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "building_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Building Id"
            }
          },
          {
            "name": "office_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Office Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offices/": {
      "get": {
        "tags": [
          "offices"
        ],
        "summary": "Get All Offices",
        "operationId": "get_all_offices_offices__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/parkings/": {
      "post": {
        "tags": [
          "parkings"
        ],
        "summary": "Park Vehicle",
        "operationId": "park_vehicle_parkings__post",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParkRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "parkings"
        ],
        "summary": "Get Parkings",
        "operationId": "get_parkings_parkings__get",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/parkings/{numberplate}/unpark": {
      "patch": {
        "tags": [
          "parkings"
        ],
        "summary": "Unpark Vehicle",
        "operationId": "unpark_vehicle_parkings__numberplate__unpark_patch",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "numberplate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Numberplate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/billing": {
      "get": {
        "tags": [
          "billing"
        ],
        "summary": "Get Bill",
        "operationId": "get_bill_billing_get",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 12,
              "minimum": 1,
              "title": "Month"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1970,
              "title": "Year"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Get Health",
        "operationId": "get_health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddBuildingRequestDTO": {
        "properties": {
          "buildingName": {
            "type": "string",
            "title": "Buildingname"
          }
        },
        "type": "object",
        "required": [
          "buildingName"
        ],
        "title": "AddBuildingRequestDTO"
      },
      "AddFloorRequestDTO": {
        "properties": {
          "floor_number": {
            "type": "integer",
            "title": "Floor Number"
          }
        },
        "type": "object",
        "required": [
          "floor_number"
        ],
        "title": "AddFloorRequestDTO"
      },
      "AddOfficeRequestDTO": {
        "properties": {
          "office_name": {
            "type": "string",
            "title": "Office Name"
          },
          "floor_number": {
            "type": "integer",
            "title": "Floor Number"
          }
        },
        "type": "object",
        "required": [
          "office_name",
          "floor_number"
        ],
        "title": "AddOfficeRequestDTO"
      },
      "AddVehicleRequestDTO": {
        "properties": {
          "numberplate": {
            "type": "string",
            "title": "Numberplate"
          },
          "type": {
            "type": "integer",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "numberplate",
          "type"
        ],
        "title": "AddVehicleRequestDTO"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginDTO": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 10,
            "minLength": 2,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginDTO"
      },
      "ParkRequestDTO": {
        "properties": {
          "numberplate": {
            "type": "string",
            "title": "Numberplate"
          }
        },
        "type": "object",
        "required": [
          "numberplate"
        ],
        "title": "ParkRequestDTO"
      },
      "RegisterDTO": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "maxLength": 10,
            "minLength": 3,
            "title": "Name"
          },
          "officeId": {
            "type": "string",
            "title": "Officeid"
          },
          "password": {
            "type": "string",
            "maxLength": 10,
            "minLength": 3,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "name",
          "officeId",
          "password"
        ],
        "title": "RegisterDTO"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}